requirements.txt
streamlit>=1.24.0
pdfplumber
python-docx
spacy
google-generativeai
sqlalchemy
pandas
requests
python-dotenv
